testing

+++ src/controller/Position6DOF.cpp
@@ -100,11 +100,10 @@ void Position6DOF::predictNextPosition( Position6DOF olderPosition, long int tim
 	long int timediff = this->timestamp - olderPosition.getTimestamp();
 	double xDiff = this->position[0] - olderPosition.getPosition()[0];
 	double yDiff = this->position[1] - olderPosition.getPosition()[1];
-	double rate = timeInFuture / timediff;
+	double rate = (double) timeInFuture / timediff;
 	double xNew = this->position[0] + xDiff*rate;
 	double yNew = this->position[1] + yDiff*rate;
 	double zOld = this->position[2];	// Value should be ignored
-
 	this->position[0] = xNew;
 	this->position[1] = yNew;
 	this->position[2] = zOld;
